[#213] Refactor: Category Executor 패턴을 단일 CategoryService로 통합
