+++ src/main/java/soon/fridgely/domain/category/event/CategoryEventListener.java
@@ -7,7 +7,7 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
-import soon.fridgely.domain.category.service.CategoryAppender;
+import soon.fridgely.domain.category.service.CategoryService;
 import soon.fridgely.domain.refrigerator.event.RefrigeratorCreatedEvent;
 
 @Slf4j
@@ -15,7 +15,7 @@ import soon.fridgely.domain.refrigerator.event.RefrigeratorCreatedEvent;
 @Component
 public class CategoryEventListener {
 
-    private final CategoryAppender categoryAppender;
+    private final CategoryService categoryService;
 
     /**
      * MemberService.register()의 @Transactional이 성공적으로 커밋된 이후에만 실행.
@@ -24,7 +24,7 @@ public class CategoryEventListener {
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void handleRefrigeratorCreated(RefrigeratorCreatedEvent event) {
         try {
-            categoryAppender.appendDefaultCategories(event.toKey());
+            categoryService.appendDefaultCategories(event.toKey());
         } catch (Exception e) {
             log.error("[CategoryEvent] 기본 카테고리 생성 실패 (RefrigeratorId={}, MemberId={})", event.refrigeratorId(), event.memberId(), e);
         }
